Add generated persistence model, EF migration, and byte spawn coords for Castle Siege
- Regenerate BasicModel and EntityFramework Model classes for the new
  Castle Siege data-model types (were missing from the PR, which broke the
  Persistence build via GameConfiguration.Generated referencing them).
- Add EF Core migration AddCastleSiege (new data tables + GameConfiguration
  config column) and update the model snapshot.
- Change CastleSiegeNpcDefinition.SpawnX/SpawnY from short to byte to match
  the map-coordinate convention used across the codebase (maps to smallint).

+++ src/DataModel/Configuration/CastleSiegeNpcDefinition.cs
@@ -35,12 +35,12 @@ public partial class CastleSiegeNpcDefinition
     /// <summary>
     /// Gets or sets the X coordinate of the NPC's spawn position.
     /// </summary>
-    public short SpawnX { get; set; }
+    public byte SpawnX { get; set; }
 
     /// <summary>
     /// Gets or sets the Y coordinate of the NPC's spawn position.
     /// </summary>
-    public short SpawnY { get; set; }
+    public byte SpawnY { get; set; }
 
     /// <summary>
     /// Gets or sets the facing direction of the NPC at spawn.
